group shift_left and shift_right
